Fixed folder capitalization for Linux-based servers, fixed rare permissions bug
Some Linux-based servers experienced a bug with the ACF data folder not being read due to the capitalization of the folders, this has been fixed. The other bug was a rare one - under certain conditions, when the damage permissions are all loading, the map default can load FIRST and THEN it checked for which permission mode was the global default. This would cause an issue where, on load, the map default would be set and then it would immediately get set to the default permission mode of "none". This has also been fixed.